# back-end/src/shared/infra/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TypeOrm_Postgre API",
    "description": "...",
    "version": "1.0.0"
  },
  "host": "localhost:3002",
  "basePath": "/",
  "tags": [
    {
      "name": "Authenticated Routes",
      "description": "Routes authenticated by Bearer Token JWT.",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "Users_auth",
      "description": "Endpoints for adding pets to the directory"
    },
    {
      "name": "Products_auth",
      "description": "Endpoints for updating pet information"
    },
    {
      "name": "Categories_auth",
      "description": "Endpoints for getting information about pets"
    },
    {
      "name": "Open Routes",
      "description": "Routes of Class Users",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "Users",
      "description": "Endpoints for adding pets to the directory"
    },
    {
      "name": "Products",
      "description": "Endpoints for updating pet information"
    },
    {
      "name": "Categories",
      "description": "Endpoints for getting information about pets"
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "description": "Please, insert in value the word 'Bearer' before token provided in signin route.\nExample: Bearer TOKEN"
    }
  },
  "paths": {
    "/categories": {
      "post": {
        "tags": [
          "Categories_auth"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Canecas Brancas2"
                },
                "description": {
                  "type": "string",
                  "example": ""
                },
                "defaultPrice": {
                  "type": "number",
                  "example": 50
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": [
              {}
            ]
          }
        ]
      }
    },
    "/categories/{id}": {
      "put": {
        "tags": [
          "Categories_auth"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Category ID.",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": ""
                },
                "description": {
                  "type": "string",
                  "example": ""
                },
                "defaultPrice": {
                  "type": "number",
                  "example": 0
                },
                "activated": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": [
              {}
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Categories_auth"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Category ID.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": [
              {}
            ]
          }
        ]
      }
    },
    "/products": {
      "post": {
        "tags": [
          "Products_auth"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": ""
                },
                "description": {
                  "type": "string",
                  "example": ""
                },
                "category_id": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": [
              {}
            ]
          }
        ]
      }
    },
    "/changeProduct/{id}": {
      "put": {
        "tags": [
          "Products_auth"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID.",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": ""
                },
                "description": {
                  "type": "string",
                  "example": ""
                },
                "category_id": {
                  "type": "string",
                  "example": ""
                },
                "activated": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": [
              {}
            ]
          }
        ]
      }
    },
    "/users": {
      "post": {
        "tags": [
          "Users_auth"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Lucca"
                },
                "username": {
                  "type": "string",
                  "example": "Luccazx12"
                },
                "password": {
                  "type": "string",
                  "example": "teste123@"
                },
                "email": {
                  "type": "string",
                  "example": "whatdogs1222@gmail.com"
                },
                "role_id": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": [
              {}
            ]
          }
        ]
      }
    },
    "/allUsers": {
      "get": {
        "tags": [
          "Users_auth"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearerAuth": [
              {}
            ]
          }
        ]
      }
    },
    "/users/roles": {
      "get": {
        "tags": [
          "Users_auth"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearerAuth": [
              {}
            ]
          }
        ]
      }
    },
    "/users/search/{username}": {
      "get": {
        "tags": [
          "Users_auth"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "User username.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": [
              {}
            ]
          }
        ]
      }
    },
    "/allProducts": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/allCategories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/user/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User username."
          }
        ],
        "responses": {}
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Lucca"
                },
                "username": {
                  "type": "string",
                  "example": "Luccazx12"
                },
                "password": {
                  "type": "string",
                  "example": "teste123@"
                },
                "email": {
                  "type": "string",
                  "example": "whatdogs1222@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/signin": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "mario-lucca@hotmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "1234"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    }
  }
}